test: consolidate browser tests into the vtbrowser module
Port the one regression guard internal/browsertest had that vtbrowser
lacked — TestBrowser_signalAttrClassStyleApply, the real-browser proof
that Signal.Attr/Class/Style bindings actually apply (the #128
colon-syntax fix) — then delete internal/browsertest wholesale.

internal/browsertest was the older //go:build browser harness, never
wired into CI (no job, script, or Makefile referenced it). vtbrowser is
the CI-run harness (its own module, dedicated Browser Tests job), so the
two were redundant; the only unique coverage is preserved by the port.
